fixed form parsing for fields defined as sub-objects and boolean parsing.
finally got around to fixing sub-objects so the TEST form is parsed
correctly now.  warning, only works for one level of sub-objects. we'll
fix that when we actually have a form in production using it. also
removed one redundant test and added boolean field to TEST form for
better coverage of different field types.  also added test for fieldsToHtml.

fixed boolean fields that were returning false when given the wrong
field type (string instead of 0/1). returning null if we don't see a
0/1.