just exit when config has been updated rather than try to reload. WIP
Reloading the dates and parts of the app that need refreshing got
somewhat messy, so just exiting.  Since sentinel is monitored by a
supervisor anyway it works out.  Also there are some memory issues we
are working out that became apparent when using VM with small memory
allocation, so this seemed like the most stable route to having
the process in a good state.