fixed bug in modal updating and rewrote update facility bulk edit.
added `isBulkUpdateDone` and another listener to call it so we know when
to close a modal that is still going through an update, we check the
docs in the changes feed against the docs in the bulk update response.
couldn't figure out a better way to do this.  otherwise the modal will
close on any new updates on changes feed.

replaced backbone functionality using new modal/bulkedit stuff. still
need to remove backbone stuff, I don't think it's used anymore. created
`updateFacility` and `getFacilityDesc` utility functions and template.