include non-form fields on ORPT record.
refactored a bit with formatDate helper.

With some forms like ORPT (patient registration), we add additional data to
it based on other form submissions.  Form data from other reports is used to
create these fields and it is useful to show these new fields in the data
records screen/render even though they are not defined in the form.