refactored changes feed init for data records.
created data_records.onDataRecordInit() to setup changes feed since that
feed only belongs to the screen/module.  Also so I can handle the error
state better and use a modal instead of an alert which is kind of
annoying because it blocks the browser event loop. In progress to Moving
error handling to modal window within data record module.