added default app_settings file.
Default app_settings values. This file will get merged with the actual
`app_settings` property on the design doc.

If you add a new value to kanso.json settings_schema then you should update
this file also.  If you forget to update this then new defauls won't show up
in settings until it gets saved in the dashboard. This can be an issue if
you have a new feature that expects a setting default value.