fixes for #392, include sms message in export for incoming records.
The main issue here was that `null` was not being set on the incoming
unstructured message so it wasn't include in the export.

We have two types of messages to include then exporting unstructured
messages.  Tasks are for outgoing messages and for incoming unstructured
unstructured messages we use the original message that is saved on the
sms message object.

Modified and added a test also.