added support for `send_day` on scheduled message config.
This allows you to specify Monday-Sunday when configuring a reminder.
This is calculated by adding the offset then setting to the closest
send_day.  Note, there is potential to schedule a message earlier then
the offset, for now it is preferred to set the day close as possible to
the offset as opposed to always after the offset.

If we choose to do only after the offset in the future it would be a
pretty simple while loop like so:

```
if (send_day) {
    var test = due.clone().day(send_day);
    while(due.day() !== test.day()) {
        due.add('1 day');
    }
}
```