towards #406; replaced patient_id search with extended fulltext search.
added new fulltext via in `lib/fulltext.js` for data records. Indexing
all fields that are strings or numbers except: 'type', 'form', 'from',
'_rev', 'refid', and 'id'.  Note this does not include `_id` you can
still do a search for `_id:<uuid>` and get one record, I left this
for a future use case where we might want to shared a link to a specific
document for editing.

This support wildcards and keywords like AND, so you can compose
searches using a colon and field prefix like: `year:2013 AND month:9`.
If you are a district admin then all search are restricted to your
district.

Removed patient_id related views since couchdb-lucene will index them
and various other fields as well.

Removed the title on the page in favor of making the search box use the
limited real-estate in the page header. It's redundant anyway since we
have the highlighted tab.

This was originally requested by Kemri because their caregiver data is
submitted via SMS so only exists on the record.  They need a way to find
what patients a caregiver is responsible for.  So the default keys when
searching without a field prefix are: patient_id, patient_name,
caregiver_name and caregiver_phone.  Ideally we should not configure the
index like this since every project will use different field names. But
for now will do the the trick for the Kemri requirment.  It doesn't
hurt anything and keeps current support for searching by patient_id. To
search other fields you need to prefix the search term with the field
name and a colon.  I tried to access the design doc from within the
fulltext view (couchdb-lucene) so that we might be able to specify what
fields are indexed via settings, but it was not possible.  This might be
a limitation and could probably be solved later by replacing our
currently settings web API with a node process that could handle this.

The interface should be improved as this continues to evolve, to allow
for searching by a field without needed to know the key string on the
doc.  I will release as-is and we can continue to think about ways
to make it better and get feedback.