fixed bug in contact name rendering in activity rows. #408
This looks like a bug not only because it's a duplicated line but also
undoes the work in the preceding block for a contact name on a clinic
(lowest level contact info), hence never rendering a reporting unit
contact name even when available. Pretty sure we want to show contact
names when possible here.