added contact_display_short setting  #408
This allows for custom rendering on the first column in the activity
field.  It's a direct mapping on the `related_entities` property on the
record.  Not the most user friendly and flexible solution but the
quickest.  Allows you to show your records using any single property
within that object. Default remains `contact.name` or town name so
existing configurations should not be affected.

Also increased the size of the contact-info div use to display the name
by 5%. It was useful real estate that was not being used by the other
columns. Let me know if you think that is a potential issue.