limit outgoing message processing to 25 records at a time. #410
In some cases if the gateway is left off for a long period, many
messages will get queued up. This can cause problems on the gateway
since there is a limit to how much data it can process.  So if 500
records have outbound messages then it will take 20 minutes to clear
that message queue for those records if the gateway is set to poll every
minute. Remember a record can have many outgoing messages, so this
isn't a limit on outgoing messages, it's a limit on the amount of
records the view will return.  You can override this by passing in a
`limit` query param on the URL.