0.3.0-beta.31: refactored tasks_pending views/lists, fixed bug in limit
added new endpoint `/add/limit/:limit` for specifying limit on outgoing
messages.

`limit` can't be passed in via query param, it needs to be parsed
by the rewrite otherwise the default always overrides.  The previous
commit added the limit of 25 docs to `/add`.  This is necessary because
the gateway needs to parse the pending messages and do a callback
request to update the docs.  This could be a rather large request so it
needs a limit to prevent the gateway from blowing up or acting
unpredictably.

Also did some refactoring to move the logic of what documents should be
parsed to the view, since the limit applies to the view.  Otherwise a
the view would return docs that the list would never return and our
pending queue would never get cleared.  So fixed a bug in the previous
commit.

Also added `tasks_sent` view for future use to calculate the amount of
sent messages.