minor bugfix on export query params and UX adjustments
Added new lines to rows on XML export so it's easier to inspect.

Made English CSV export default. SpreadsheetML can be a little buggy
because we're using HTML entities (not valid XML).

Disabled default month value in exports screen since record count does
not reflect the export row totals displayed.  It's probably better UX to
have the user set the date knowing they are doing something than having a
default that doesn't make sense with the totals on the screen and having
to guess why that is.

TODO default export start month to subtract setting and calculate based
on the the date value.