use logger.debug() in sms parser instead of logger.log().  #431
This keeps log output more quiet by default since loglevel property in
kanso.json is set to 'info'.

For now keeping this as-is in kanso.json, which has not front end editor
other than futon. So it will likely only be modified when pushing a new
design doc.  Eventually it should move to garden style `app_settings`
property so a new code push is not required.