Modified expected `message` value throughtout configs. #486
This is a breaking change.  Now expecting an array of objects with
"content" and "locale" properties, both properties are required.

You should see a warning if you are still using the old structure:

"Message property should be an array. Upgrade your schedule
configuration to 0.4 compatibility."

This change affects schedules, validations, registrations and patient
reports.  It only applies to the configuration structure, not the
actually document structure, so message locale on a config is resolved
when it gets applied to a document.  The document does not store all the
message data but only the message in the expected locale.

The locale is required for the time being, if you do not include locale
your messages will be ignored.  This should be an easy change later if
necessary.