added support for `?replace=1` query parameter to app-settings #283
Modified upload forms configuration to use replace.  So now the upload
will replace app_settings.forms instead of extending it.

Modified app-settings rewrites to specify PUT method on updates.
@gareth have a review here to make sure this restriction doesn't break
configuration updates, thought it made sense to explicitly define the
method.

Removed objectpath rewrite on updates in app-settings package, it never
really existed and is unnecessary, you just include the object with the
property you want to update in the body of the request.