allow registration transition to run when doc has errors. #480
There are validations done on the couchdb side before sentinel processes
the document.  These validations can add errors to the doc. For example
when required:true is set on a field in the form definition and that
field is not present after parsing the message.

I'm not sure the original intention for the registration transition to
skip documents with errors but removing it for now until further reason
presents itself.

Up until now most of the forms I have been dealing with left the fields
optional, so most of the validation was done on the backend as opposed
to muvuku or the program doing the form submission. So I never came
across this issue.