fixed message parsing bug, support locale on smsparser.isCompact. #555
not totally happy with this patch, couldn't decide whether isCompact
should use the locale on the msg (since it can optionally be an object),
or pass in as a separate parameter. Chose the latter.

The reason being we modify the message (remove form code) in
smsparser.parse before we do the parsing, so we should probably avoid
using the object since it gets modified. So thinking to make addl patch
to modify smsparser.isCompact to only accept a string for a message.