fixed bug with download forms settings button.
The button was based on the server side/first render data, not the
uploaded data.  So made it part of renderInstalledForms loop/function.
Also it was showing while forms were uploading. Sometimes that can take
a long time, and people might think they need to click on it.  So hide
download button during uploading and re-render and link uploaded form
data.