support for adding new forms via JSON (ODK Collect Support #715)
Somewhat of a major refactor of the updates module.  Took liberty and
removed dead code, mainly `updates.updateRelated` and anything only in
that code path. Also refactored or removed any related tests.

Added new updates function called `add` that calls `add_sms` or
`add_json` depending on how the request submits data.  The `add_sms`
path is used if the request includes a `message` parameter, otherwise
we assume it's JSON.  The new `add_json` update function will return an
error if the JSON cannot be parsed or a form cannnot be found.