Fix for #925: tag HTTP requests with their update target
This adds a new parameter for HTTP requests made through the
`HttpWrapper` service - `updateTarget`.  Current options for
`updateTarget` are `filter`, `left` and `right`, referring to different
sections of the UI which an HTTP request might update on successful
completion.

When the user navigates to different states, the `stateChangeStart`
listener selectively cancels pending HTTP requests depending on the old
and new states.