Toward medic/medic-webapp#903.
Dependency added: jszip.

Provide basic support for exporting a zip file containing concatenated server
log files. Currently, this is a single human-readable Markdown file in a zip
container. In the future, we will split the Markdown file up in to several
package-specific directories and service-specific files. We'll also provide
options to limit the export scope, and/or increase the number of exported log
segments.

This changeset depends on the following Medic OS change:
  https://github.com/browndav/medic-os/commit/3f79c2f405ec5e6ead84d7adaa8b962d38ccf1b2

Not quite ready to merge yet. This needs a test or two, as well as a change to
the `format` parameter that medic/medic-webapp is invoking us with.