Move login template from `public/` to `templates/` to avoid confusion
When located in `public/`, the login index page can be accessed at
`http://<hostname>:<port>/login`, but if accessed at this URL it does not have
translations and does not work.  This is very confusing if you access URL when
trying to find the actual working login page for the app.

Fix for https://github.com/medic/medic-webapp/issues/1078