cleaned up kanso.json settings, removed defaults.
Issue: https://github.com/medic/medic-webapp/issues/741

Removed most of the string type defaults in kanso.json settings_schema.
This should cause a little less headache when trying to
remove/clear/disable settings that are not used.

We still have a bug in the json.edit module on the dashboard json editor
on array and integer types, I'm not sure how to make them empty or not
required.

The net effect is not very different because even without a default
value the editor still used the empty string by default for string
types.