added support for outgoing_deny_list conf in default_response transition.
outgoing_deny_list is a new configuration that is a comma separated list
of strings. If a string in that list matches the beginning of the phone
we are setting up a response to then this transition is ignored.  The
transition will never fire because the filter will fail on those
documents and those message will be left without an auto-reply.

Issue: https://github.com/medic/medic-webapp/issues/750