Use SettingsP more widely. Reinstate settings cache. Issue: #1507
This fixes a few places where we wanted a promise but wrapped the Settings service which then calls the SettingsP service. By having these places call SettingsP directly this issue goes away. This means the cache can then be stored rather than recreated every time which means it actually works as a cache.