refactored outgoing deny list config handling. Issue #750
Outgoing deny list will now be applied to schedules and responses
generated on reports.  So instead of seeing `scheduled` or `pending`
state on a message you will see `denied`.

Moved `isOutgoingAllowed`, `_isMessageFromGateway` and related tests
into utils library so schedule creation and report validation and
response messages can make use of it.

Calling `messages.addMessage` or `uitls.addScheduledMessage` now
transparently checks the outgoing deny list and assigns a `denied` state
when appropriate.

I removed some odd logic in `addScheduledMessage` because it's not used
or documented anywhere and is error prone.  There was a check for
`doc.muted` field when creating a scheduled message.  I looked through
the code and fairly certain this property is not set or used by
anything.