Make radio button clicking easier and more obvious on mobile
Now, when clicking on a radio button, the background of the button will become a
lovely LG-orange, and the text white.  This occurs immediately the control is
touched, which gives the user much more visible feedback than the previous light
blue that was used.

Also the touchable vertical area of radio buttons is increased to make them
harder to miss.

Issue: #1630