Temporary fix tour bug. Issue: #933
bootstrap-tour has a bug in its latest tag with tours that contain orphan
steps. This has been fixed but only in master; there is not yet a release.

I've forked the repo and created a v0.10.3 tag that contains v0.10.2 plus
this one fix which we can use until they release a real version.