Clean up loading indicators
On some pages (notably the tasks page) the loading indicator doesn't
work properly. This is due to offloading work in $timeout calls, which
gives Angular a chance to do work between those calls. To fix this I
moved the loading state change to occur after the work done in $timeout
calls.