fix initial replication bug so all docs are synced. Issue #1695
DBSync was using `live:true` on all replication requests so that will
only get new changes.  It does not initiate replication from update_seq
0. This could be the cause of the perceived "app never finishes
installing" problem.  Trying to think of bad repercussions of this
change but not coming up with much.

I added an extra replication that is `live:false` and only makes the
continuous replication after it completes.  There is no `complete` event
for `live:true` replication requests becuase it's designed to never
complete.