Automatic Ddoc extraction #36
This feature supports our ability to have muiltiple DDocs being served out of
the same Kanso project.

On the `medic-webapp` side, all extra Ddocs should be JSON files (with their
correct `_id` etc) inside the `ddocs/compiled` directory. This is then picked
up by Kanso as an attachment and pushed to Couch.

When `medic-api` runs it will run this code. This code looks through all
attached Ddocs, and re-uploads them if it thinks it should.