Support weakly-validated ETag headers
This fixes a repeated spurious "Update available" message on page
load/reload. The version detection relies on an ETag header but
does not correctly discard the `W/` prefix for weak ETags, leading
to a presumed version mismatch.

This doesn't appear to be covered by unit tests, and probably should
be split out to an ETag parser module one way or another so it can
be unit tested without lots of screaming and gnashing of teeth.