Don't allow .container-fluid to scroll
When scrolling downward very quickly, it's possible to reach the
end of `#contacts-list` and accidentally scroll vertically in
the ancestor `.container-fluid`. This causes the contacts list to visually
shrink and be replaced from the bottom with a dull, flat, useless gray box.
I've been able to reproduce this on desktop and mobile; this trivial CSS tweak
appears to fix the problem.

Signed-off-by: David Brown <david@medicmobile.org>