Failed queries can be retried
Increase reliabilty by allowing Pouch/CouchDB queries to be retried. In theory
nothing should go wrong locally, but if you're calling out to Couch (ie admin)
a lot of data gets pulled down, and so network reliabilty comes into play.

This could be made a lot better by making it clear that it failed, but at least
for now it just removes the loading indicator and allows you to tap and try
again.

Issue: #2054