Bind least specific data first, then more specific
Our data structure doesn't match 1:1 to our XML forms structure, so we
allow a data value to be bound to any XML decendents that match it.
If the data value is an object, then we find the XML decendent that
matches it and that is our new scope.
The change orders so that values are bound first, so that more specific
data structures (objects) get to bind over less specific ones.

Issue: #2146