Add pagionation support to db-object-widget
Usess select2's automatic pagination magic to load more results as you
scroll. Will probably we a bit weird because we sort results by name
on the client side, which then means it is almost certainly true that
subsequent sets of data will not be in the "right" order. This could
possibly be fixed by looking at sorting views in Couch.

Issue: #2161