Rework filters for performance, technical debt, and usability. Issue: #1866 #1863
Move the search filters down from being the InboxCtrl scope to being in the scope where they're applied. For example, the form type filter is now in the ReportsCtrl scope. This improves our onload performance by only loading the filters as they are needed. It reduces the number of elements in the DOM and consequently the number of watchers on each page. It also cleans up the InboxCtrl by removing tab specific code.