folds FacilityLegacy back into Facility
We might as well with this new caching strategy, it doesn't make it
any slower.

I'm keeping the warn in there because the idea of us downloading all
person documents down and having that invalidate if any one of them
changes is something we need to get rid of in the future.

Issue: #2209, #2210, #2211, #2218