ported a patch from `diy` branch that was lost.
cf4c3ff fixed bug where default response did not respect locale value

Not sure if this is necessary for DIY/v0.4 compatibility. I know there
is a project or two that use this feature.  It allows you to support
reporters in multiple languages based on a field named `locale` set in
the form or a query parameter `locale` set on the gateway.

Issue: https://github.com/medic/medic-webapp/issues/2410