Replaced backlog processor. Issue: medic/medic-webapp#2521 (#81)
Replaced the broken `last_valid_seq` view with a doc with id `sentinel-meta-data` that records the last seq the backlog processor got to. This starts at 0 so existing sites will start processing again but transitions are designed to be able to be run multiple times so other than a slight performance hit it should be fine.