Either search or scope to type (#2690)
* Scope contacts search #2339

Search defaults (and is constrained) to the direct level below the
user's facility type. a CHP will only see families for example.
Admins are excluded and see everything.

* Either search or scope to type #2339

Search is either a raw text search of contacts you can access, or its
all contacts that are the type one "below" your facility. E.g., for a
CHW that would be all families, for a branch manager that would be all
CHW areas, etc.