Use couchdb 2.0 on travis
One of the changes made in this commit is to switch the travis configuration
from using project type (`language`) of `node_js` to `java`.  This allows us to
take control directly of which JDK is being used to build the project.  This
is necessary because otherwise Travis will silently change the JRE used for node
projects.  This led to the following error for some builds:

	The command "node --stack_size=10000 `which grunt` ci" exited with 1.
	$ cat nohup.out
	seleniumProcess.pid: undefined
	events.js:160
	      throw er; // Unhandled 'error' event
	      ^

	Error: spawn java ENOENT
	    at exports._errnoException (util.js:1012:11)
	    at Process.ChildProcess._handle.onexit (internal/child_process.js:182:32)
	    at onErrorNT (internal/child_process.js:348:16)
	    at _combinedTickCallback (internal/process/next_tick.js:74:11)
	    at process._tickCallback (internal/process/next_tick.js:98:9)
	    at Module.runMain (module.js:577:11)
	    at run (bootstrap_node.js:352:7)
	    at startup (bootstrap_node.js:144:9)
	    at bootstrap_node.js:467:3

	The command "cat nohup.out" exited with 0.

I have tested with `jdk` values of `oraclejdk7` and `oraclejdk8`, and only the
former currently works.  I expect this will change with future versions of
webdriver-manager/selenium/protractor.

Issue: #2933