Generates patient_id on patients, auto-determine id length
* Generate patient_id on patients

New sentinel transition that adds a patient_id to each person document that
does not already contain one. This allows patients created via XForms to be used
via SMS. It will also allow CHWs to be referred to via SMS in the future if a
partner wishes to do so.

* Auto-lengthen ids generated

To support an increase in ids, and to remove a production issue where sentinel
runs out of generatable ids, we now auto-lengthen the patient ids we generate.

We attempt to create 5 randomly generated ids, and if all of them are already
used by patients we up the length of ids we're generating by 1 and try again.
This will auto-increase as they are used up to a maximum of 14, which is in the
area of a million billion ids, so I'm sure it's fine for now™.

* Drop support for id_format

Dropping support for this setting allows us to more easily support ids
that auto-lengthen when required. id_format was not used by any existing
partners, with the exception of fixing the aforementioned issue with Sentinel
running out of ids.

Tickets:
 - medic/medic-webapp#3166
 - medic/medic-webapp#3000
 - medic/medic-webapp#3230