Improve trigger param parsing
Errors in trigger parsing now propagate correctly through the callback stack
instead of being lost. We also now catch situations where JSON inside the
string is malformed, instead of presuming being unable to parse it means
it's an comma-delimited string.

Node: We actually *do* support raw JSON, at least in theory. Other areas
may read kanso.json and think that app_settings is invalid because that
value should be a string, and break. For example, where we auto-generate
config editing. So we are adding support for it, but on the dl for now.

We then also still support parsing json from a string, and this
will continue to be our official documentation about it for now.