Introducing system transitions
System transitions are run by default, as opposed to normal ones which
must be configured to be run first.

This is for transitions that are maintaining application / data
correctness in some way, where we would very much expect everyone to
run these transitions.

NB: configurers can still disable them if they really want to, using
    explicit disabling in config.