Updated release note example
The previous example was accurate but not representative of the way the rule was being used in configs. Changed to make it easier to know what to search and replace in app_settings configuration. Also, using `\d` in config needs to be escaped in JSON so preferable to use `[0-9]` to avoid errors.