Don't clear selected if loading specific task
Because we load the livelist and the task report in parallel it's
highly likely the list won't contain the ID on page load. So instead of
clearing if the item isn't in the livelist, only clear if we haven't
been given an ID to select. This is consistent with other tabs.

medic/medic-webapp#3090