app_setting translation import / export fixes
In medic/medic-webapp#3127 support is added for moving message
translation out of app_settings into the standard translation files.

Here, we add support for this change in import and export functions.

Translation export no longer exports when the translation is actually in
the translation file. This stop empty lines with a key but no value
appearing in translation files.

Translation import ignores translation lines when importing into
app_settings, so as to not break on exports generated prior to this
change.

medic/medic-webapp#3350