Fix app_setting translation import / export (#3350)
In medic/medic-webapp#3127 support was added to app_settings to allow for
translations to be declared as keys referencing translations in our main
translation files, instead of directly storing the translations alongside config
in app_settings.

Here, we are modifying the translation import and export functions to support
and understand this change.

For export, we no longer export translations in situations when the translation
is just a key reference to the translation file. This stops lines with a key but
no value being generated on export.

For import, we ignores translation lines with a key but no value, so as to not
break on exports generated prior to the change noted above.