Attaching grunt-contrib-watch to a tag
I had to fork grunt-contrib-watch to fix a bug where it watched more
directories than it should on MacOS, causing grunt to fire when it
shouldn't.

Previously this was attached to the branch. I have now tagged it and
attached it to that instead. This unfortunately doesn't stop npm from
constantly re-requesting it from git: the only way to do that would be
for it to be in NPM.