Fixes daterangepicker on mobile
Our custom mobile daterangepicker was broken by an update in the
daterangepicker library at some point. This modifies our custom code
to no longer hide the picker and lets the library do that automatically
via a new 'autoApply' option which selects as soon as the end date is
selected. This also applies to the desktop but it's an improvement
there as well I think.

I also removed a bunch of translations which are no longer used.

medic/medic-webapp#3467