Adds a listener to Enketo to notify on change
Enketo has an event that fires when the form content has been updated
by the user. This change leverages that so that the "may lost content"
warning dialog is only shown when the user has actually changed a
value.

This fixes a bug where the warning is shown when the form failed to
render, and also a bug where the warning is shown when the form has no
fields.

medic/medic-webapp#1601
medic/medic-webapp#3045