id generator now stores id length in a doc
Instead of beginning at length 5, and storing any increase in this
value in memory (and thus restarting at 5 upon restart) this is now
stored in the `shortcode-id-length` document.

This lets tech leads and others create or edit this document to set
their desired minimum length. Additionally, if the id length
increments, it won't decrement and generate smaller ids on restart,
which could be confusing.

medic/medic-webapp#3505