Added minimum length constraint
As noted in the comment, this is because the cache size has been worked
out to be roughtly good for 5-7 digits (1,110,000 ids), and if you set
it smaller you could potentially get into infinite loops and other
issues.

If we want to allow smaller ids at some point we'd just need to revisit
these assumptions.