Hides form title in CSS instead of removing from form
The upload_xform script was removing the title from the xform
definition so it wouldn't show in the UI. This broken medic-collect
integration which relies on the title being present.

This change leaves the title in the form. It is replaced during
render with the translated version from the doc. This is hidden on
mobile where the title is displayed in the title bar instead.

medic/medic-webapp#3390