Reduces select2 update binding scope
When the selected db object is changed in a select2 widget we
attempt to bind to other fields defined by that object. This
allows form configurations that can interrogate other fields
on the object.

The bug was that when binding the code matched all the fields in
all the selects, rather than just the fields in this repeat
section.

medic/medic-webapp#3430