Replication key falls back to contact
When we detect that the patient report references a patient but we
can't find them by the given ID the replication key falls back to
the contact ID. This means that restricted users that can
replicate the CHW can replicate the report.

This allows for the CHW and supervisors to see the report and
resubmit with the correct patient id.

medic/medic-webapp#3437