Hide contact when not found
Before this change when a document had a contact or parent that
wasn't found in the database we defaulted to showing the UUID
because that's all the information we had available. This is a
common situation for restricted users who don't replicate the whole
tree.

Now the parent is hidden altogether so the only names shown are of
contacts you can view.

medic/medic-webapp#3752
medic/medic-webapp#3769