Don't write remote checkpoints
By default pouchdb writes checkpoints locally and remotely when
a document is replicated. The checkpoint is so the next replication
can pick up where this one left off. For us this means lots of
checkpoints being replicated over slow connections.

This change keeps the checkpoints in the local database but removes
them on the remote database.

medic/medic-webapp#3644