Removing legacy mockPromise usage
We used to have to use a 3rd party promise library. This is no longer
the case. Remove the now-pointless redirection.

Note that it does still have some minor value, in being a utility to
correctly reject or resolve (err, payload) dynamically, so we have
kept it for just that and renamed it from 'mockPromise' to 'promise'.