Make getRegistrations usage consistent
Previously getRegistrations returned CouchDB rows (with an embedded doc
property), and the callers extracted the doc out in various different
places, including sometimes trying to do it twice!

Instead, always return just the list of docs so it's done in one place.

Incidentally this also makes the documentation about the function in
the comment above it correct.

medic/medic-webapp#3838