Update font to Noto and sizes to rems
Updates the font to Noto and bundles the font with the app so all clients will
look the same. Increases the base font size from 14px to 16px for readability.
Introduces 5 standard font sizes in rems that can be applied anywhere and result
in the same output. Refactors HTML to be more semantic to use standardised
classes (h1, h2, h3, etc). Reworks the CSS to use px instead of em which is more
reliable as it's not cumulative.

medic/medic-webapp#3597