Throw an error when document not found
The lineage model generator was returning an empty model when the
document with the given ID was not found. This caused uncaught
errors and made it less obvious to the user what was going wrong.
This change throws a 404 error which can be interpreted by the
caller and displayed to the user in some useful way.

medic/medic-webapp#3718