Mark required fields when adding restricted user
The facility and contact are required only when adding a restricted
user, so this change adds an asterix when the type is set to
restricted.

Also adds a css class "required" for use on the form-group which
adds an asterix after the label which makes it easier to keep the
style consistent througout the app.

medic/medic-webapp#3561