Inner docs should have correct lineages
1. Siblings (in the case of a family form, the head of household)
already get their parent lineage mostly correct, because they attach the
doc as the parent, (see the if (child) block in
prepareAndAttachSiblingDocs()) which then later on (see the promiseChain
in the same function) get loaded from CouchDB with its whole hierarchy.
However, they still need to extract the lineage to clean it up

2. Repeated docs preperation was moved until *after*
prepareAndAttachSiblings was run, so that the doc they prepare against
has its parents loaded correctly (the same state the siblings rely on)

NB: this is really gross. It relies on non-obvious ordering. I've added
comments, but we should consider reworking this

medic/medic-webapp#3840