Check for userCtx cookie before bootstrapping
Previously we checked for the cookie after bootstrapping which
meant if the redirect to the login page was cancelled, for example
by pressing "esc", then the user could access the offline data
without a cookie or session. The navigation can still be cancelled
but the app won't be bootstrapped so the data cannot be easily
accessed.

This change also fixes how we remove cookies so it actually works.

medic/medic-webapp#3239