Update resource icons to inline svgs
The ResourceIcons service and resourceIcons filter now detect if
the icon is an svg and inline the xml if it is. This means we can
now apply css styles to the svg which is important, for example,
when we want to use a coloured icon for clinic in the list and the
same icon in white in the action bar.

medic/medic-webapp#3945