Allow SMS sent to other people based on bool_expr
* Refactoring: more cleanly expose silenceReminders

The Registration transition currently calls handleReport, but only to
silenceReminders. Instead of having it call a function that also does
other things (and because of passed vars, only does that for
registrations), just call through to that one function.

This does mean there is more replicated code, but IMO it's cleaner

* Allow SMS sent to other people based on bool_expr

Two main changes:
 - Allow for you to set the recipient phone number to an arbitrary JS
   expression, given the doc as context
 - Allow for you to decide whether or not to send out an SMS based on
   an arbitrary JS expression, given the doc as context

This lets you, for example, send an SMS to a parent contact based on
some values in the incoming report.

* Fixing zero-length lineage docs

Types. I miss thee.

The old code was broken somehow, passing around the nano object instead
of a document. This new code does what you'd expect. I've also added
a test.

medic/medic-webapp#3412