Refactor buildHydratedDoc to be more readable
Get rid of currentParent and currentStub: they are the same. If we want
them to be different (if we want to only map to a stored stub lineage
and not one that grows organically out of attaching non-stub parents) we
would need to soft-clone them to get a new Object.

Re-ordered how we find out starting point to make clear the decision
between contact types and data records.

Re-ordered the while loop to make clear our walking of the lineage and
how we're attaching things.

Which in turn made clear some potential bugs to me, which I added Errors
for. FWIW I'm not convinced the first Error (attaching where there is a
non-existant stub) is actually an error. It could be that the stored
lineage we have is out of date and the linage we are accumulating from
attaching hydrated parents is more correct.

---

Which in turn is interesting: we are pulling hydratedParents based on
the hierarchy in the document. If that hierarchy is out of date the
current code will both ignore it and not ignore it: it will ignore it
and pull whatever ids are in the list, it will not ignore it if we use a
stub, and in geneal it could be wrong because it's replacing one
hydrated document's parent with the incorrect parent.

For example, if you have a non-hydrated doc with doc.a.b.c.d as parents,
but at some point a's parent was changed to e, you will do the
following:

 - Pull a, b, c and d as parents to hydrate
 - Attach doc to a, a to b, b to c and c to d
 - In the old code: ignoring that the real parent of a is now e.
   In the new code, throwing an error because b._id !== a.parent._id

We are probably not dealing with this case right now, but it feels right
to have the guards in there for the future (including the invisible
future where a random TL does this in PROD without mentioning it)

---

Some tests fail. I think these are part of the discussion.